Changed example in `oldof` use case for globals
Even if this use case is desired, people may vote against it because of the example provided. We do not want to encourage contracts related to time measurements. We would get a "Schroedinger precondition" that is satisfied when contracts are not runtime-checked (because functions now execute faster and the intervals between time measurements are smaller) but is not satisfied when runtime-checking is disabled because now functions do additional runtime instructions and take longer. Thus, only enabling/disabling runtime checking would affect program correctness (if measured by violated contracts).